Create list_gen_to_sort.py
This is a little program that runs a while loop to create a list where the numbers are transformed to an int type there for we can manage it as a hybrid list as it were made by hand and then take that int data and convert in to str data, so we can do a sort method to organize all the data in the list. 

I made this little project because I was trying the sort method on a list with str and int, and it didn't work because I didn't manage operation between types, so i decide to create something that gives the sort output without the need of me putting the numbers as a str input. Then I saw someone in making endless input function and I combine those two to make something more complex, but that gives me my original desire output.